+++ crates/bevy_webview_wry/src/webview/load_webview/ipc.rs
@@ -2,21 +2,29 @@ use bevy::ecs::system::SystemParam;
 use bevy::log::error;
 use bevy::prelude::{Entity, Res};
 use bevy_flurx_ipc::ipc_commands::{IpcCommand, IpcCommands, Payload};
-use bevy_flurx_ipc::prelude::{IpcRawEvent, IpcRawEventBody, IpcRawEvents};
+use bevy_flurx_ipc::prelude::{IpcTriggerMessage, IpcTriggerSender};
 use serde::Deserialize;
 use wry::WebViewBuilder;
 
 #[derive(SystemParam)]
 pub(crate) struct IpcHandlerParams<'w> {
     ipc_commands: Res<'w, IpcCommands>,
-    ipc_raw_events: Res<'w, IpcRawEvents>,
+    ipc_raw_events: Res<'w, IpcTriggerSender>,
 }
 
 #[derive(Deserialize)]
 #[serde(tag = "type", content = "message")]
-enum IpcMessage {
+enum WebviewIpcMessage {
     Command(Payload),
-    Event(IpcRawEventBody),
+    Event(IpcTriggerBody),
+}
+
+#[derive(Deserialize)]
+struct IpcTriggerBody {
+    /// event id
+    pub event_id: String,
+    /// The serialized main body of the event sent from the webview.
+    pub payload: String,
 }
 
 impl IpcHandlerParams<'_> {
@@ -26,20 +34,21 @@ impl IpcHandlerParams<'_> {
         builder: WebViewBuilder<'a>,
     ) -> WebViewBuilder<'a> {
         let ipc_commands = self.ipc_commands.clone();
-        let ipc_raw_events = self.ipc_raw_events.clone();
+        let sender = self.ipc_raw_events.clone();
 
         builder.with_ipc_handler(move |request| {
-            match serde_json::from_str::<IpcMessage>(request.body()) {
-                Ok(IpcMessage::Command(payload)) => {
+            match serde_json::from_str::<WebviewIpcMessage>(request.body()) {
+                Ok(WebviewIpcMessage::Command(payload)) => {
                     ipc_commands.push(IpcCommand {
                         entity: webview_entity,
                         payload,
                     });
                 }
-                Ok(IpcMessage::Event(payload)) => {
-                    ipc_raw_events.push(IpcRawEvent {
-                        webview_entity,
-                        body: payload,
+                Ok(WebviewIpcMessage::Event(body)) => {
+                    sender.send(IpcTriggerMessage {
+                        target: Some(webview_entity),
+                        event_id: body.event_id,
+                        payload: body.payload,
                     });
                 }
                 Err(e) => {